+++ src/baas/src/secbaas/community/core/repository/resource_key/_protocol.py
@@ -1,18 +1,5 @@
-from typing import Protocol, runtime_checkable
+"""Resource key repository protocol — re-exported from api.api_gateway."""
 
-from ._record import ResourceKeyRecord
+from secbaas.community.api.api_gateway import ResourceKeyRecord, ResourceKeyRepository
 
-
-@runtime_checkable
-class ResourceKeyRepository(Protocol):
-    """Resource key repository protocol."""
-
-    def get_by_resource_key_and_tenant(
-        self, resource_key: str, tenant: str
-    ) -> ResourceKeyRecord | None:
-        """Look up record by resource_key and tenant."""
-        ...
-
-    def exists_bot_mapping(self, resource_key_id: int, bot_id: str) -> bool:
-        """Check whether a mapping between resource_key_id and bot_id exists."""
-        ...
+__all__ = ["ResourceKeyRecord", "ResourceKeyRepository"]
